refactor: move `define_into_tonic_status` to `common-error` (#4095)
* chore: finish cherry-pick

* chore: remove unused code
