+++ src/common/error/Cargo.toml
@@ -10,3 +10,4 @@ workspace = true
 [dependencies]
 snafu.workspace = true
 strum.workspace = true
+tonic.workspace = true
