+++ src/flow/src/adapter/error.rs
@@ -16,12 +16,12 @@
 
 use std::any::Any;
 
+use common_error::define_into_tonic_status;
 use common_error::ext::BoxedError;
 use common_macro::stack_trace_debug;
 use common_telemetry::common_error::ext::ErrorExt;
 use common_telemetry::common_error::status_code::StatusCode;
 use datatypes::value::Value;
-use servers::define_into_tonic_status;
 use snafu::{Location, Snafu};
 
 use crate::adapter::FlowId;
