+++ src/frontend/src/error.rs
@@ -15,10 +15,10 @@
 use std::any::Any;
 
 use common_datasource::file_format::Format;
+use common_error::define_into_tonic_status;
 use common_error::ext::{BoxedError, ErrorExt};
 use common_error::status_code::StatusCode;
 use common_macro::stack_trace_debug;
-use servers::define_into_tonic_status;
 use snafu::{Location, Snafu};
 use store_api::storage::RegionNumber;
