+++ src/operator/Cargo.toml
@@ -47,7 +47,6 @@ prometheus.workspace = true
 query.workspace = true
 regex.workspace = true
 serde_json.workspace = true
-servers.workspace = true
 session.workspace = true
 snafu.workspace = true
 sql.workspace = true
